reverted it back to valley's version www 	modified:   exmmtest.map 	modified:   scroll.map 	modified:   src/lib/vgmsnd/stdtype.h 	modified:   src/lib/vgmsnd/vgmSnd.c 	modified:   src/lib/vgmsnd/vgmSnd.h 	modified:   vgmtest.exe

+++ src/lib/vgmsnd/stdtype.h
@@ -9,8 +9,8 @@ typedef uint8_t	UINT8;
 typedef  int8_t	 INT8;
 typedef uint16_t	UINT16;
 typedef  int16_t	 INT16;
-typedef uint32_t	UINT32;
-typedef  int32_t	 INT32;
+typedef signed long	UINT32;
+typedef  unsigned long	 INT32;
 typedef uint64_t	UINT64;
 typedef  int64_t	 INT64;
